Split index & react entries
