--- index.html
@@ -1,61 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<link rel="stylesheet" href="main.css">
-</head>
-
-<body>
-	<main>
-		<section class="display">
-		</section>
-		<section class="controller">
-			<h2>Intersection Observer</h2>
-			<p>This proof of concept uses the <a
-					href="https://developer.mozilla.org/en-US/docs/Web/API/Intersection_Observer_API">Interesection Observer</a>
-				to
-				determine which of the 10,000 elements on the left are in view and should be rendered.</p>
-			<p>As the user scrolls, the intersection observer callback receives only those elements with changed
-				<code>isInteresecting</code> status
-			</p>
-			<p>Zooming in and out is performant and works properly, because layout is performed by the browser engine and not
-				javascript</p>
-
-			<div class="observer-option">
-				<label for="threshold"><code>threshold</code></label>
-				<input value="1.0" id="threshold-value" class="value-display" oninput=onSliderChange()>
-				<input value="1.0" oninput=onSliderInput(this) onchange=onSliderChange() type="range" id="threshold"
-					name="threshold" min="0" max="1.0" step="0.1">
-			</div>
-
-			<div class="observer-option">
-				<label for="margin"><code>margin</code></label>
-				<input value="0" id="margin-value" class="value-display" oninput=onSliderChange()>%
-				<input value="0" oninput=onSliderInput(this) onchange=onSliderChange() type="range" id="margin" name="margin"
-					min="0" max="100" step="1">
-			</div>
-
-			<div class="observer-option">
-				<label for="number"><code>number</code></label>
-				<input value="1000" id="number-value" class="value-display" oninput=onSliderChange()>
-				<input value="1000" oninput=onSliderInput(this) onchange=onSliderChange() type="range" id="number" name="number"
-					min="10" max="50000" step="10">
-			</div>
-
-			<div class="observer-option">
-				<button id="toggle-view">Toggle list / grid view</button>
-			</div>
-
-			<p id="rendered-number"></p>
-
-			<p><a href="./react.html">See it done with React</a></p>
-			<p><a href="https://github.com/rendall/intersection-observer-grid">Github repo</a></p>
-
-		</section>
-
-		<script src="main.js"></script>
-	</main>
-</body>
-
-</html>
