--- main.css
@@ -1,88 +1,0 @@
-* {
-  margin: 0;
-  box-sizing: border-box;
-}
-
-main {
-  font-family: arial;
-  display: flex;
-}
-main section {
-  width: 50vw;
-}
-main section.display {
-  border: solid black 1px;
-  display: flex;
-  flex-wrap: wrap;
-  height: 100vh;
-  width: 50%;
-  overflow-y: scroll;
-}
-main section.display h2 {
-  flex-basis: 100%;
-  margin: 1rem;
-  font-size: 3rem;
-}
-main section.display article.cell {
-  border: solid red 1px;
-  color: gray;
-  height: 10rem;
-  margin: 1rem;
-  overflow: hidden;
-  position: relative;
-  width: 10rem;
-}
-main section.display article.cell img.image-display {
-  position: absolute;
-  top: 0;
-  left: 0;
-  display: none;
-}
-main section.display article.cell p {
-  background: none;
-}
-main section.display article.cell p.number-display {
-  color: red;
-  display: none;
-  font-size: 3rem;
-  position: absolute;
-  text-align: center;
-  top: 50%;
-  transform: translateY(-50%);
-  width: 100%;
-}
-main section.display article.cell.is-view, main section.display article.cell.is-react {
-  color: red;
-  font-weight: bold;
-}
-main section.display article.cell.is-view p.number-display,
-main section.display article.cell.is-view img.image-display, main section.display article.cell.is-react p.number-display,
-main section.display article.cell.is-react img.image-display {
-  display: block;
-}
-main section.display.list-view article.cell {
-  width: 100%;
-}
-main section.controller {
-  padding: 1rem;
-}
-main section.controller > h2,
-main section.controller p {
-  margin-bottom: 1rem;
-}
-main section.controller input.value-display {
-  width: 3rem;
-}
-main section.controller .observer-option {
-  display: flex;
-  flex-direction: row;
-  margin-bottom: 1rem;
-}
-main section.controller .observer-option input {
-  margin: 0 1rem;
-}
-main section.controller .observer-option input[type=range] {
-  flex-grow: 1;
-}
-
-/*# sourceMappingURL=main.css.map */
