+++ package.json
@@ -1,29 +1,35 @@
 {
   "scripts": {
+    "prebuild": "rimraf ./dist",
     "build": "webpack --mode=production --node-env=production",
-    "build:dev": "npm start && webpack --mode=development",
-    "build:prod": "npm start && webpack --mode=production --node-env=production",
+    "build:dev": "webpack --mode=development",
+    "build:prod": "webpack --mode=production --node-env=production",
     "build:sass": "npx sass main.scss:main.css",
     "start": "npx http-server -a localhost",
-    "watch": "webpack --watch"
+    "watch": "webpack --watch",
+    "serve": "webpack serve"
   },
   "devDependencies": {
     "@types/react": "^17.0.13",
     "@types/react-dom": "^17.0.8",
     "@webpack-cli/generators": "^2.2.0",
+    "html-webpack-plugin": "^5.3.2",
+    "mini-css-extract-plugin": "^2.2.0",
     "prettier": "^2.3.2",
-    "sass": "^1.35.1",
+    "rimraf": "^3.0.2",
+    "sass": "^1.37.5",
     "sass-loader": "^12.1.0",
-    "ts-loader": "^9.2.3",
+    "ts-loader": "^9.2.5",
     "typescript": "^4.3.5",
-    "webpack": "^5.43.0",
-    "webpack-cli": "^4.7.2"
+    "webpack": "^5.50.0",
+    "webpack-cli": "^4.7.2",
+    "webpack-dev-server": "^3.11.2"
   },
   "dependencies": {
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
   "version": "1.0.0",
-  "description": "React intersection observer PoC",
-  "name": "react-intersection-observer-poc"
+  "description": "My webpack project",
+  "name": "my-webpack-project"
 }
