+++ src/react-app.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { App } from './components/App';
+import {App} from './components/App'
 const domContainer = document.querySelector('#react-app');
 ReactDOM.render(React.createElement(App), domContainer);
