+++ src/react.html
@@ -8,7 +8,7 @@
 
 <body>
 	<main id="react-app"></main>
-	<script src="dist/main.js"></script>
+	<script src="react.js"></script>
 </body>
 
 </html>
