+++ webpack.config.js
@@ -1,15 +1,25 @@
-// Generated using webpack-cli https://github.com/webpack/webpack-cli
-
 const path = require("path");
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 
 const isProduction = process.env.NODE_ENV == "production";
 
 const config = {
-  entry: "./src/index.ts",
+  entry: { react:"./src/react-app.ts", index:"./src/index.ts" },
   output: {
     path: path.resolve(__dirname, "dist"),
   },
+  devServer: {
+    open: true,
+    host: "localhost",
+  },
   plugins: [
+    new HtmlWebpackPlugin({
+      template: "./src/react.html",
+    }),
+
+    new MiniCssExtractPlugin(),
+
     // Add your plugins here
     // Learn more about plugins from https://webpack.js.org/configuration/plugins/
   ],
